Start master learning objectives list
Started creating a master list of learning objectives in the index.md
file to facilitate organization of the lessons later (and to provide
learners with a bird's eye view of the entire lesson before starting).

I was doing this work offline, so I didn't have access to material from
previous SWC python lessons, which probably covers similar topics much
better.